refactoring all JS files in samples that containing React components to JSX
