+++ samples/charts/AreaChartConfig.jsx
@@ -17,68 +17,68 @@
  */
 
 import React from 'react';
-import { ChartWrapper } from './ChartWrapper';
+import { ChartWrapper } from './ChartWrapper.jsx';
 import './css/Table.css';
 import VizG from '../../src/VizG.jsx';
 
-export default class LineChartConfigSample extends React.Component {
+export default class AreaChartConfigSample extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
             data: [[1, 10, 23, 'piston'], [1, 20, 34, 'rotary']],
             data2: [[1, 10, 23, 'piston']],
-            timer: 0,
+            timer: 0
         };
-        this.interval_id = null;
+
         this.metadata = {
-            names: ['rpm', 'torque', 'horsepower', 'EngineType', 'weight'],
-            types: ['linear', 'linear', 'linear', 'ordinal', 'linear'],
+            names: ['rpm', 'torque', 'horsepower', 'EngineType'],
+            types: ['linear', 'linear', 'linear', 'ordinal']
         };
-        this.lineChartConfig = {
+
+        this.interval_id = null;
+        this.areaChartConfig = {
             x: 'rpm',
-            charts: [{ type: 'line', y: 'torque', color: 'EngineType', colorDomain: ['', '', 'piston'] }],
+            charts: [{ type: 'area', y: 'torque', color: 'EngineType', colorDomain: ['', '', 'piston'] }],
             maxLength: 7,
             width: 700,
             height: 450,
-            axisLabelColor: 'red',
-            tickLabelColor: 'red',
-            // animation: true
+
         };
 
-        this.singleLineChartConfig = {
+        this.stackedAreaChartConfig = {
             x: 'rpm',
-            charts: [
-                { type: 'line', y: 'horsepower', fill: '#2ca02c' },
-                { type: 'line', y: 'torque', fill: '#ff7f0e' },
-            ],
+            charts: [{ type: 'area', y: 'torque', color: 'EngineType', colorDomain: ['', '', 'piston'], mode: 'stacked' }],
             maxLength: 7,
             width: 700,
             height: 450,
-            // animation: true
+
         };
 
-        this.sparkLineConfig = {
+        this.singleAreaChartConfig = {
             x: 'rpm',
-            charts: [{ type: 'spark-area', y: 'horsepower', fill: '#0fd8e2' }],
-            maxLength: 30,
-            width: 120,
-            height: 10,
-            // animation: true
+            charts: [{ type: 'area', y: 'horsepower', fill: '#2ca02c' }, { type: 'area', y: 'torque', fill: '#ff7f0e' }],
+            maxLength: 7,
+            width: 700,
+            height: 450,
+
         };
     }
+
 
     componentDidMount() {
         this.interval_id = setInterval(() => {
             this.setState({
                 data: [
                     [this.state.timer, this.state.timer === 20 ? null : Math.random() * 100, 10, 'piston'],
-                    [this.state.timer, Math.random() * 100, 10, 'rotary'],
+                    [this.state.timer, Math.random() * 100, 10, 'rotary']
                 ],
                 data2: [
-                    [this.state.timer, Math.random() * 100, Math.random() * 100, 'rotary'],
+
+                    [this.state.timer, Math.random() * 100, Math.random() * 100, 'rotary']
                 ],
-                timer: this.state.timer + 1,
+                timer: this.state.timer + 1
             });
+
         }, 500);
     }
 
@@ -89,40 +89,65 @@ export default class LineChartConfigSample extends React.Component {
     render() {
         return (
             <div>
-                <center><h1>Line Chart Config Samples</h1></center>
-                <ChartWrapper title="Group MultiLine Chart Sample" chart="line" media actionBar={false}>
-                    <VizG config={this.lineChartConfig} metadata={this.metadata} data={this.state.data} />
-                    <br /><br />
-                    <div style={{ display: 'block' }}>
+                <center><h1>Area Chart Config Samples</h1></center>
+                <ChartWrapper title="Group Area Chart Sample" chart="line" media={true} actionBar={false}>
+                    <VizG config={this.areaChartConfig} metadata={this.metadata} data={this.state.data} />
+                    <br />
+                    <div>
+
                         <pre>
+                            <p>Configuration: </p>
+
                             {'{\n' +
                                 '\tx: \'rpm\',\n' +
-                                '\tcharts: [\n\t    { type: \'line\', y: \'torque\', color: \'EngineType\',colorDomain:[\'\',\'\',\'piston\']}\n\t],\n' +
+                                '\tcharts: [\n\t    { type: \'area\', y: \'torque\', color: \'EngineType\',colorDomain:[\'\',\'\',\'piston\']}\n\t],\n' +
                                 '\tmaxLength: 7,\n' +
                                 '\twidth: 700,\n' +
                                 '\theight: 450,\n' +
                                 '\tanimation:true\n}'
+
                             }
                         </pre>
                     </div>
                 </ChartWrapper>
-                <ChartWrapper title="Multi Line Chart Sample" chart="line" media actionBar={false}>
-                    <VizG config={this.singleLineChartConfig} metadata={this.metadata} data={this.state.data2} />
+                <ChartWrapper title="Multi Area Chart Sample" chart="line" media={true} actionBar={false}>
+                    <VizG config={this.singleAreaChartConfig} metadata={this.metadata} data={this.state.data2} />
                     <br />
                     <div>
                         <pre>
+                            <p>Configuration: </p>
                             {'{\n' +
                                 '\tx: \'rpm\',\n' +
-                                '\tcharts: [\n\t    { type: \'line\', y: \'horsepower\', fill:\'#2ca02c\'}\n\t    { type: \'line\', y: \'torque\', fill:\'#ff7f0e\'}\n\t],\n' +
+                                '\tcharts: [\n\t    { type: \'area\', y: \'horsepower\', fill:\'#2ca02c\'}\n\t    { type: \'area\', y: \'torque\', fill:\'#ff7f0e\'}\n\t],\n' +
                                 '\tmaxLength: 7,\n' +
                                 '\twidth: 700,\n' +
                                 '\theight: 450,\n' +
                                 '\tanimation:true\n}'
+
                             }
                         </pre>
                     </div>
                 </ChartWrapper>
-                
+                <ChartWrapper title="Group Area Chart Sample stacked" chart="line" media={true} actionBar={false}>
+                    <VizG config={this.stackedAreaChartConfig} metadata={this.metadata} data={this.state.data} />
+                    <br />
+                    <div>
+
+                        <pre>
+                            <p>Configuration: </p>
+
+                            {'{\n' +
+                                '\tx: \'rpm\',\n' +
+                                '\tcharts: [\n\t    { type: \'area\', y: \'torque\', color: \'EngineType\',mode:\'stacked\',colorDomain:[\'\',\'\',\'piston\']}\n\t],\n' +
+                                '\tmaxLength: 7,\n' +
+                                '\twidth: 700,\n' +
+                                '\theight: 450,\n' +
+                                '\tanimation:true\n}'
+
+                            }
+                        </pre>
+                    </div>
+                </ChartWrapper>
                 <ChartWrapper title="Sample Data Set" chart="line">
                     <div>
                         <pre>
@@ -148,20 +173,6 @@ export default class LineChartConfigSample extends React.Component {
                                         <th>Description</th>
                                     </tr>
                                     <tr>
-                                        <td>blah</td>
-                                        <td>blah</td>
-                                        <td>
-                                            
-                                            <span>
-                                                <VizG
-                                                    config={this.sparkLineConfig}
-                                                    metadata={this.metadata}
-                                                    data={this.state.data2}
-                                                />
-                                            </span>
-                                        </td>
-                                    </tr>
-                                    <tr>
                                         <td>x</td>
                                         <td>string</td>
                                         <td>independent axis</td>
@@ -172,30 +183,11 @@ export default class LineChartConfigSample extends React.Component {
                                         <td>Charts to be plotted</td>
                                     </tr>
                                     <tr>
-                                        <td>yAxisTickAngle</td>
-                                        <td>int</td>
-                                        <td>tick angle value of yAxis ticks | default : 0</td>
-                                    </tr>
-                                    <tr>
-                                        <td>xAxisTickAngle</td>
-                                        <td>int</td>
-                                        <td>tick angle value of xAxis ticks | default : 0</td>
-                                    </tr>
-                                    <tr>
-                                        <td>axisLabelColor</td>
-                                        <td>String(color)</td>
-                                        <td>Label Color of the axis label</td>
-                                    </tr>
-                                    <tr>
-                                        <td>tickLabelColor</td>
-                                        <td>String(color)</td>
-                                        <td>Label Color of the tick label</td>
-                                    </tr>
-                                    <tr>
                                         <td>maxLength</td>
                                         <td>int</td>
                                         <td>Maximum length of the dataSet displayed</td>
                                     </tr>
+
                                     <tr>
                                         <td>height</td>
                                         <td>int</td>
@@ -230,8 +222,9 @@ export default class LineChartConfigSample extends React.Component {
                                     <tr>
                                         <td>mode</td>
                                         <td>string</td>
-                                        <td>type of the line curve (see <a href="https://github.com/d3/d3-shape#curves">d3-documentation</a> for possible curve types)</td>
+                                        <td>whether the chart should be stacked or not (default:null)</td>
                                     </tr>
+
                                     <tr>
                                         <td>colorScale</td>
                                         <td>string | Array(string)</td>
