+++ samples/charts/PieChartsConfigSample.jsx
@@ -15,78 +15,68 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+
 import React from 'react';
-import { ChartWrapper } from './ChartWrapper';
+import { ChartWrapper } from './ChartWrapper.jsx';
 import './css/Table.css';
 import VizG from '../../src/VizG.jsx';
 
-export default class BarChartConfigSample extends React.Component {
+export default class LineChartConfigSample extends React.Component {
+
     constructor(props) {
         super(props);
         this.state = {
             data: [[1, 10, 23, 'piston'], [1, 20, 34, 'rotary']],
             data2: [[1, 10, 23, 'piston']],
-            timer: 0
+            timer: 0,
         };
+
+        this.interval_jd = null;
 
         this.metadata = {
-            names: ['rpm', 'torque', 'horsepower', 'EngineType'],
-            types: ['linear', 'linear', 'linear', 'ordinal']
+            names: ['rpm', 'torque', 'horsepower', 'EngineType', 'weight'],
+            types: ['linear', 'linear', 'linear', 'ordinal', 'linear'],
         };
 
-        this.interval_id = null;
-
-        this.barChartConfig = {
-            x: 'rpm',
-            charts: [{ type: 'bar', y: 'torque', color: 'EngineType', colorDomain: ['', '', 'piston'] }],
-            maxLength: 20,
-            width: 700,
-            height: 450,
-
+        this.lineChartConfig = {
+            charts: [{ type: 'arc', x: 'torque', color: 'EngineType', mode: 'donut' }],
+            width: 300,
+            height: 300,
+            legendOrientation: 'top',
         };
 
-        this.barHorizontalChartConfig = {
-            x: 'rpm',
-            charts: [{ type: 'bar', y: 'torque', color: 'EngineType', colorDomain: ['', '', 'piston'], orientation: 'left' }],
-            maxLength: 20,
-            width: 700,
-            height: 450,
-            // animation: false
+        this.configPie = {
+            charts: [{ type: 'arc', x: 'torque', color: 'EngineType', mode: 'pie' }],
+            width: 300,
+            height: 300,
         };
 
-        this.stackedBarChartConfig = {
-            x: 'rpm',
-            charts: [{ type: 'bar', y: 'torque', color: 'EngineType', colorDomain: ['', '', 'piston'], mode: 'stacked' }],
-            maxLength: 20,
-            width: 700,
-            height: 450,
-            // animation: true
-        };
-
-        this.singleBarChartConfig = {
-            x: 'rpm',
-            charts: [{ type: 'bar', y: 'horsepower', fill: '#2ca02c' }, { type: 'bar', y: 'torque', fill: '#ff7f0e' }],
-            maxLength: 20,
-            width: 700,
-            height: 450,
-            // animation: true
+        this.configT = {
+            charts: [{ type: 'arc', x: 'torque', color: 'EngineType', colorScale: ['steelblue', '#80ccff'] }],
+            tooltip: { enabled: false },
+            legend: false,
+            percentage: true,
+            colorScale: ['steelblue', '#80ccff'],
+            width: 300,
+            height: 300,
         };
     }
 
     componentDidMount() {
-        this.interval_id = setInterval(() => {
+        this.interval_jd = setInterval(() => {
             this.setState({
                 data: [
                     [this.state.timer, this.state.timer === 20 ? null : Math.random() * 100, 10, 'piston'],
-                    [this.state.timer, Math.random() * 100, 10, 'rotary']
+                    [this.state.timer, Math.random() * 100, 10, 'rotary'],
+                    [this.state.timer, this.state.timer === 20 ? null : Math.random() * 100, 10, 'piston2'],
+                    [this.state.timer, Math.random() * 100, 10, 'rotary2'],
                 ],
                 data2: [
 
-                    [this.state.timer, Math.random() * 100, Math.random() * 100, 'rotary']
+                    [this.state.timer, Math.random() * 100, Math.random() * 100, 'rotary'],
                 ],
-                timer: this.state.timer + 1
+                timer: this.state.timer + 1,
             });
-
         }, 500);
     }
 
@@ -97,87 +87,54 @@ export default class BarChartConfigSample extends React.Component {
     render() {
         return (
             <div>
-                <center><h1>Bar Chart Config Samples</h1></center>
-                <ChartWrapper title="Group MultiLine Chart Sample" chart="line" media={true} actionBar={false}>
-                    <VizG config={this.barChartConfig} metadata={this.metadata} data={this.state.data} />
-                    <br />
-                    <div>
-
+                <center><h1>Line Chart Config Samples</h1></center>
+                <ChartWrapper title="Donut Chart Sample" chart="line" media actionBar={false}>
+                    <VizG config={this.lineChartConfig} metadata={this.metadata} data={this.state.data} />
+                    <br /><br />
+                    <div style={{ display: 'block' }}>
                         <pre>
-                            <p>Configuration: </p>
-
-                            {'{\n' +
-                                '\tx: \'rpm\',\n' +
-                                '\tcharts: [\n\t    { type: \'bar\', y: \'torque\', color: \'EngineType\',colorDomain:[\'\',\'\',\'piston\']}\n\t],\n' +
-                                '\tmaxLength: 7,\n' +
-                                '\twidth: 700,\n' +
-                                '\theight: 450,\n' +
-                                '\tanimation:true\n}'
+                            {
+                                '{\n' +
+                                '\tcharts : [{type:"arc", x:"torque", color:"EngineType", mode:"donut"}],\n' +
+                                '\twidth: 300,\n' +
+                                '\theight: 250\n' +
+                                '}'
 
                             }
                         </pre>
                     </div>
                 </ChartWrapper>
-                <ChartWrapper title="Multi Line Chart Sample" chart="line" media={true} actionBar={false}>
-                    <VizG config={this.singleBarChartConfig} metadata={this.metadata} data={this.state.data2} />
+                <ChartWrapper title="Pie Chart Sample" chart="line" media actionBar={false}>
+                    <VizG config={this.configPie} metadata={this.metadata} data={this.state.data} />
                     <br />
                     <div>
                         <pre>
-                            <p>Configuration: </p>
-                            {'{\n' +
-                                '\tx: \'rpm\',\n' +
-                                '\tcharts: [\n\t    { type: \'bar\', y: \'horsepower\', fill:\'#2ca02c\'}\n\t    { type: \'bar\', y: \'torque\', fill:\'#ff7f0e\'}\n\t],\n' +
-                                '\tmaxLength: 7,\n' +
-                                '\twidth: 700,\n' +
-                                '\theight: 450,\n' +
-                                '\tanimation:true\n}'
+                            {
+                                '{\n' +
+                                '\tcharts : [{type: "arc",  x : "torque", color : "EngineType", mode: "pie"}],\n' +
+                                '\twidth: 400,\n' +
+                                '\theight: 300\n' +
+                                '}'
+                            }
+                        </pre>
+                    </div>
+                </ChartWrapper>
+                <ChartWrapper title="Donut Chart Sample" chart="line" media actionBar={false}>
+                    <VizG config={this.configT} metadata={this.metadata} data={this.state.data} />
+                    <br /><br />
+                    <div style={{ display: 'block' }}>
+                        <pre>
+                            {
+                                '{\n' +
+                                '\tcharts : [{type:"arc", x:"torque", color:"EngineType", mode:"donut"}],\n' +
+                                '\twidth: 300,\n' +
+                                '\theight: 250\n' +
+                                '}'
 
                             }
                         </pre>
                     </div>
                 </ChartWrapper>
-                <ChartWrapper title="Group MultiLine Chart Sample Horizontal" chart="line" media={true} actionBar={false}>
-                    <VizG config={this.barHorizontalChartConfig} metadata={this.metadata} data={this.state.data} />
-                    <br />
-                    <div>
-
-                        <pre>
-                            <p>Configuration: </p>
-
-                            {'{\n' +
-                                '\tx: \'rpm\',\n' +
-                                '\tcharts: [\n\t    { type: \'bar\', y: \'torque\', color: \'EngineType\',orientation:\'left\',colorDomain:[\'\',\'\',\'piston\']}\n\t],\n' +
-                                '\tmaxLength: 7,\n' +
-                                '\twidth: 700,\n' +
-                                '\theight: 450,\n' +
-                                '\tanimation:true\n}'
-
-                            }
-                        </pre>
-                    </div>
-                </ChartWrapper>
-
-                <ChartWrapper title="Group MultiLine Chart Sample stacked" chart="line" media={true} actionBar={false}>
-                    <VizG config={this.stackedBarChartConfig} metadata={this.metadata} data={this.state.data} />
-                    <br />
-                    <div>
-
-                        <pre>
-                            <p>Configuration: </p>
-
-                            {'{\n' +
-                                '\tx: \'rpm\',\n' +
-                                '\tcharts: [\n\t    { type: \'bar\', y: \'torque\', color: \'EngineType\',mode:\'stacked\',colorDomain:[\'\',\'\',\'piston\']}\n\t],\n' +
-                                '\tmaxLength: 7,\n' +
-                                '\twidth: 700,\n' +
-                                '\theight: 450,\n' +
-                                '\tanimation:true\n}'
-
-                            }
-                        </pre>
-                    </div>
-                </ChartWrapper>
-
                 <ChartWrapper title="Sample Data Set" chart="line">
                     <div>
                         <pre>
@@ -202,10 +159,21 @@ export default class BarChartConfigSample extends React.Component {
                                         <th>Type</th>
                                         <th>Description</th>
                                     </tr>
+
                                     <tr>
                                         <td>x</td>
                                         <td>string</td>
                                         <td>independent axis</td>
+                                    </tr>
+                                    <tr>
+                                        <td>labelColor</td>
+                                        <td>string(color)</td>
+                                        <td>label color of the percentage mode</td>
+                                    </tr>
+                                    <tr>
+                                        <td>innerRadius</td>
+                                        <td>int(pixel)</td>
+                                        <td>innerRadius in pixels for the chart in donut mode or percentage mode</td>
                                     </tr>
                                     <tr>
                                         <td>charts</td>
@@ -222,8 +190,8 @@ export default class BarChartConfigSample extends React.Component {
                                         <td>int</td>
                                         <td>Height of the chart in pixels</td>
                                     </tr>
-
                                 </tbody>
+
                             </table>
                             <br />
                             <p>Chart Object Properties</p>
@@ -252,18 +220,14 @@ export default class BarChartConfigSample extends React.Component {
                                     <tr>
                                         <td>mode</td>
                                         <td>string</td>
-                                        <td>whether the chart should be stacked or not (default:null)</td>
-                                    </tr>
-                                    <tr>
-                                        <td>orientation</td>
-                                        <td>string</td>
-                                        <td>orientation of the chart (default:vertical)</td>
+                                        <td>type of the line curve (see <a href="https://github.com/d3/d3-shape#curves">d3-documentation</a> for possible curve types)</td>
                                     </tr>
                                     <tr>
                                         <td>colorScale</td>
                                         <td>string | Array(string)</td>
                                         <td>color set to use in the charts for predefined colors check <a
-                                            href="https://github.com/d3/d3-3.x-api-reference/blob/master/Ordinal-Scales.md#categorical-colors">d3-documentation</a></td>
+                                            href="https://github.com/d3/d3-3.x-api-reference/blob/master/Ordinal-Scales.md#categorical-colors"
+                                        >d3-documentation</a></td>
                                     </tr>
                                     <tr>
                                         <td>colorDomain</td>
