+++ samples/charts/ScatterPlotConfigSample.jsx
@@ -17,7 +17,7 @@
  */
 
 import React from 'react';
-import { ChartWrapper } from './ChartWrapper';
+import { ChartWrapper } from './ChartWrapper.jsx';
 import './css/Table.css';
 import VizG from '../../src/VizG.jsx';
