+++ samples/index.js
@@ -22,19 +22,19 @@ import {
     BrowserRouter as Router,
     Route,
 } from 'react-router-dom';
-import App from './App.jsx';
-import LineChart from './charts/LineChartsSamples';
-import BarChart from './charts/BarChartConfigSample';
-import AreaChart from './charts/AreaChartConfig';
-import ScatterPlot from './charts/ScatterPlotConfigSample';
-import PieChart from './charts/PieChartsConfigSample';
-import MapChartConfigSample from './charts/MapChartConfigSample';
-import NumberChartConfigSample from './charts/NumberChartConfigSample';
-import TableCharts from './charts/TableChartConfigSample';
 import LightBaseTheme from 'material-ui/styles/baseThemes/lightBaseTheme';
 import DarkBaseTheme from 'material-ui/styles/baseThemes/darkBaseTheme';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
+import App from './App.jsx';
+import LineChart from './charts/LineChartsSamples.jsx';
+import BarChart from './charts/BarChartConfigSample.jsx';
+import AreaChart from './charts/AreaChartConfig.jsx';
+import ScatterPlot from './charts/ScatterPlotConfigSample.jsx';
+import PieChart from './charts/PieChartsConfigSample.jsx';
+import MapChartConfigSample from './charts/MapChartConfigSample.jsx';
+import NumberChartConfigSample from './charts/NumberChartConfigSample.jsx';
+import TableCharts from './charts/TableChartConfigSample.jsx';
 
 ReactDOM.render(
     <Router>
